+++ flocker/node/_model.py
@@ -136,10 +136,15 @@ class RestartOnFailure(object):
 
         :raises ValueError: If maximum_retry_count is invalid.
         """
-        if not (self.maximum_retry_count is None or
-                self.maximum_retry_count > 0):
-            raise ValueError("maximum_retry_count must be postive or None, "
-                             "got %r" % (self.maximum_retry_count,))
+        if self.maximum_retry_count is not None:
+            if not isinstance(self.maximum_retry_count, int):
+                raise TypeError(
+                    "maximum_retry_count must be an integer or None, "
+                    "got %r" % (self.maximum_retry_count,))
+            if self.maximum_retry_count < 1:
+                raise ValueError(
+                    "maximum_retry_count must be positive, "
+                    "got %r" % (self.maximum_retry_count,))
 
 
 @attributes(["name", "image",